requirements.txt
torch
torchvision
numpy
pyyaml
tqdm
matplotlib
pandas
